new code for merge mining
