+++ src/core.cpp
@@ -5,8 +5,13 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "core.h"
+#include "coins.h"
 
 #include "util.h"
+#include "sync.h"
+
+CCriticalSection cs_main;
+CChain chainActive;
 
 std::string COutPoint::ToString() const
 {
@@ -213,11 +218,6 @@ uint64_t CTxOutCompressor::DecompressAmount(uint64_t x)
     return n;
 }
 
-uint256 CBlockHeader::GetHash() const
-{
-    return Hash(BEGIN(nVersion), END(nNonce));
-}
-
 uint256 CBlock::BuildMerkleTree() const
 {
     vMerkleTree.clear();
@@ -287,4 +287,211 @@ void CBlock::print() const
     for (unsigned int i = 0; i < vMerkleTree.size(); i++)
         LogPrintf("%s ", vMerkleTree[i].ToString());
     LogPrintf("\n");
+}
+
+bool CBlockIndex::IsSuperMajority(int minVersion, const CBlockIndex* pstart, unsigned int nRequired, unsigned int nToCheck)
+{
+  unsigned int nFound = 0;
+  for (unsigned int i = 0; i < nToCheck && nFound < nRequired && pstart != NULL; i++)
+    {
+      if (pstart->nVersion >= minVersion)
+	++nFound;
+      pstart = pstart->pprev;
+    }
+  return (nFound >= nRequired);
+}
+
+int64_t CBlockIndex::GetMedianTime() const
+{
+  AssertLockHeld(cs_main);
+  const CBlockIndex* pindex = this;
+  for (int i = 0; i < nMedianTimeSpan/2; i++)
+    {
+      if (!chainActive.Next(pindex))
+	return GetBlockTime();
+      pindex = chainActive.Next(pindex);
+    }
+  return pindex->GetMedianTimePast();
+}
+
+bool
+CAuxPow::check(const uint256& hashAuxBlock, int nChainId, const CChainParams& params) const
+{
+    if (nIndex != 0)
+        return error("AuxPow is not a generate");
+
+    if (params.StrictChainId() && parentBlock.GetChainId() == nChainId)
+        return error("Aux POW parent has our chain ID");
+
+    if (vChainMerkleBranch.size() > 30)
+        return error("Aux POW chain merkle branch too long");
+
+    // Check that the chain merkle root is in the coinbase
+    const uint256 nRootHash = CBlock::CheckMerkleBranch(hashAuxBlock, vChainMerkleBranch, nChainIndex);
+    std::vector<unsigned char> vchRootHash(nRootHash.begin(), nRootHash.end());
+    std::reverse(vchRootHash.begin(), vchRootHash.end()); // correct endian
+
+    // Check that we are in the parent block merkle tree
+    if (CBlock::CheckMerkleBranch(GetHash(), vMerkleBranch, nIndex) != parentBlock.hashMerkleRoot)
+        return error("Aux POW merkle root incorrect");
+
+    const CScript script = vin[0].scriptSig;
+
+    // Check that the same work is not submitted twice to our chain.
+    //
+
+    CScript::const_iterator pcHead =
+        std::search(script.begin(), script.end(), UBEGIN(pchMergedMiningHeader), UEND(pchMergedMiningHeader));
+
+    CScript::const_iterator pc =
+        std::search(script.begin(), script.end(), vchRootHash.begin(), vchRootHash.end());
+
+    if (pc == script.end())
+        return error("Aux POW missing chain merkle root in parent coinbase");
+
+    if (pcHead != script.end()) {
+        // Enforce only one chain merkle root by checking that a single instance of the merged
+        // mining header exists just before.
+        if (script.end() != std::search(pcHead + 1, script.end(), UBEGIN(pchMergedMiningHeader), UEND(pchMergedMiningHeader)))
+            return error("Multiple merged mining headers in coinbase");
+        if (pcHead + sizeof(pchMergedMiningHeader) != pc)
+            return error("Merged mining header is not just before chain merkle root");
+    } else {
+        // For backward compatibility.
+        // Enforce only one chain merkle root by checking that it starts early in the coinbase.
+        // 8-12 bytes are enough to encode extraNonce and nBits.
+        if (pc - script.begin() > 20)
+            return error("Aux POW chain merkle root must start in the first 20 bytes of the parent coinbase");
+    }
+
+
+    // Ensure we are at a deterministic point in the merkle leaves by hashing
+    // a nonce and our chain ID and comparing to the index.
+    pc += vchRootHash.size();
+    if (script.end() - pc < 8)
+        return error("Aux POW missing chain merkle tree size and nonce in parent coinbase");
+
+    int nSize;
+    memcpy(&nSize, &pc[0], 4);
+    const unsigned merkleHeight = vChainMerkleBranch.size();
+    if (nSize != (1 << merkleHeight))
+        return error("Aux POW merkle branch size does not match parent coinbase");
+
+    int nNonce;
+    memcpy(&nNonce, &pc[4], 4);
+
+    if (nChainIndex != getExpectedIndex(nNonce, nChainId, merkleHeight))
+        return error("Aux POW wrong index");
+
+    return true;
+}
+
+int
+CAuxPow::getExpectedIndex(int nNonce, int nChainId, unsigned h)
+{
+    // Choose a pseudo-random slot in the chain merkle tree
+    // but have it be fixed for a size/nonce/chain combination.
+    //
+    // This prevents the same work from being used twice for the
+    // same chain while reducing the chance that two chains clash
+    // for the same slot.
+
+    unsigned rand = nNonce;
+    rand = rand * 1103515245 + 12345;
+    rand += nChainId;
+    rand = rand * 1103515245 + 12345;
+
+    return rand % (1 << h);
+}
+
+FILE* OpenDiskFile(const CDiskBlockPos &pos, const char *prefix, bool fReadOnly)
+{
+    if (pos.IsNull())
+        return NULL;
+    boost::filesystem::path path = GetDataDir() / "blocks" / strprintf("%s%05u.dat", prefix, pos.nFile);
+    boost::filesystem::create_directories(path.parent_path());
+    FILE* file = fopen(path.string().c_str(), "rb+");
+    if (!file && !fReadOnly)
+        file = fopen(path.string().c_str(), "wb+");
+    int counter = 0;
+    while (!file && counter < 10) {
+        LogPrintf("Unable to open file %s\n", path.string());
+	sleep(1);
+	if (fReadOnly) {
+	  file = fopen(path.string().c_str(), "rb+");
+	}
+	else {
+	  file = fopen(path.string().c_str(), "wb+");
+	}
+        //return NULL;
+	counter ++;
+    }
+    if (!file) {
+      return NULL;
+    }
+    if (pos.nPos) {
+        if (fseek(file, pos.nPos, SEEK_SET)) {
+	  LogPrintf("Unable to seek to position %u of %s\n", pos.nPos, path.string());
+            fclose(file);
+            return NULL;
+        }
+    }
+    return file;
+}
+
+FILE* OpenBlockFile(const CDiskBlockPos &pos, bool fReadOnly) {
+    return OpenDiskFile(pos, "blk", fReadOnly);
+}
+
+bool CheckAuxPowProofOfWork(const CBlockHeader& block, const CChainParams& params)
+{
+  int algo = block.GetAlgo();
+
+  if (!block.nVersion <= 2 && params.StrictChainId() && block.GetChainId() != params.GetAuxpowChainId())
+    return error("%s : block does not have our chain ID"
+		 " (got %d, expected %d, full nVersion %d)",
+		 __func__,
+		 block.GetChainId(),
+		 params.GetAuxpowChainId(),
+		 block.nVersion);
+
+  if (!block.auxpow) {
+    if (block.IsAuxpow())
+      return error("%s : no auxpow on block with auxpow version",
+		   __func__);
+
+    if (!CheckProofOfWork(block.GetPoWHash(algo), block.nBits))
+      return error("%s : non-AUX proof of work failed", __func__);
+
+    return true;
+  }
+
+  if (!block.IsAuxpow())
+    return error("%s : auxpow on block with non-auxpow version", __func__);
+
+  if (!block.auxpow->check(block.GetHash(), block.GetChainId(), params))
+    return error("%s : AUX POW is not valid", __func__);
+
+  if(fDebug)
+    {
+      CBigNum bnTarget;
+      bnTarget.SetCompact(block.nBits);
+
+      LogPrintf("DEBUG: proof-of-work submitted  \n  parent-PoWhash: %s\n  target: %s  bits: %08x \n",
+		block.auxpow->getParentBlockPoWHash(algo).ToString().c_str(),
+		bnTarget.ToString().c_str(),
+		bnTarget.GetCompact());
+    }
+
+  if (!(algo == ALGO_SHA256D || algo == ALGO_SCRYPT) )
+    {
+      return error("%s : AUX POW is not allowed on this algo", __func__);
+    }
+
+  if (!CheckProofOfWork(block.auxpow->getParentBlockPoWHash(algo), block.nBits))
+    {
+      return error("%s : AUX proof of work failed", __func__);
+    }
+
+  return true;
 }
