+++ src/core.h
@@ -13,28 +13,15 @@
 #include "scrypt.h"
 #include <stdint.h>
 #include "hash.h"
+#include "bignum.h"
+#include "chainparams.h"
+#include "pureheader.h"
+#include "pow.h"
 
-const int NUM_ALGOS = 5;
-
-enum {
-  ALGO_SHA256D = 1,
-  ALGO_SCRYPT = 2,
-  ALGO_ARGON2 = 3,
-  ALGO_X17 = 4,
-  ALGO_LYRA2REv2 = 5
-};
+#include <boost/filesystem.hpp>
+#include <boost/filesystem/fstream.hpp>
 
 /* Use the rightmost 8 bits for standard version number, 9th bit for merge mining (todo), 10-12 th bits for POW algo, 13 th bit for update scaling factor flag */
-enum
-  {
-    BLOCK_VERSION_ALGO = (7 << 9),
-    BLOCK_VERSION_SHA256D = (1 << 9),
-    BLOCK_VERSION_SCRYPT = (2 << 9),
-    BLOCK_VERSION_ARGON2 = (3 << 9),
-    BLOCK_VERSION_X17 = (4 << 9),
-    BLOCK_VERSION_LYRA2REv2 = (5 << 9),
-    BLOCK_VERSION_UPDATE_SSF = (1 << 12)
-  };
 
 class CTransaction;
 
@@ -87,56 +74,6 @@ public:
     void SetNull() { ptx = NULL; n = (unsigned int) -1; }
     bool IsNull() const { return (ptx == NULL && n == (unsigned int) -1); }
 };
-
-/** An input of a transaction.  It contains the location of the previous
- * transaction's output that it claims and a signature that matches the
- * output's public key.
- */
-class CTxIn
-{
-public:
-    COutPoint prevout;
-    CScript scriptSig;
-    unsigned int nSequence;
-
-    CTxIn()
-    {
-        nSequence = std::numeric_limits<unsigned int>::max();
-    }
-
-    explicit CTxIn(COutPoint prevoutIn, CScript scriptSigIn=CScript(), unsigned int nSequenceIn=std::numeric_limits<unsigned int>::max());
-    CTxIn(uint256 hashPrevTx, unsigned int nOut, CScript scriptSigIn=CScript(), unsigned int nSequenceIn=std::numeric_limits<unsigned int>::max());
-
-    IMPLEMENT_SERIALIZE
-    (
-        READWRITE(prevout);
-        READWRITE(scriptSig);
-        READWRITE(nSequence);
-    )
-
-    bool IsFinal() const
-    {
-        return (nSequence == std::numeric_limits<unsigned int>::max());
-    }
-
-    friend bool operator==(const CTxIn& a, const CTxIn& b)
-    {
-        return (a.prevout   == b.prevout &&
-                a.scriptSig == b.scriptSig &&
-                a.nSequence == b.nSequence);
-    }
-
-    friend bool operator!=(const CTxIn& a, const CTxIn& b)
-    {
-        return !(a == b);
-    }
-
-    std::string ToString() const;
-    void print() const;
-};
-
-
-
 
 /** An output of a transaction.  It contains the public key that the next input
  * must be able to sign with to claim it.
@@ -193,6 +130,53 @@ public:
     }
 
     friend bool operator!=(const CTxOut& a, const CTxOut& b)
+    {
+        return !(a == b);
+    }
+
+    std::string ToString() const;
+    void print() const;
+};
+
+/** An input of a transaction.  It contains the location of the previous
+ * transaction's output that it claims and a signature that matches the
+ * output's public key.
+ */
+class CTxIn
+{
+public:
+    COutPoint prevout;
+    CScript scriptSig;
+    unsigned int nSequence;
+
+    CTxIn()
+    {
+        nSequence = std::numeric_limits<unsigned int>::max();
+    }
+
+    explicit CTxIn(COutPoint prevoutIn, CScript scriptSigIn=CScript(), unsigned int nSequenceIn=std::numeric_limits<unsigned int>::max());
+    CTxIn(uint256 hashPrevTx, unsigned int nOut, CScript scriptSigIn=CScript(), unsigned int nSequenceIn=std::numeric_limits<unsigned int>::max());
+
+    IMPLEMENT_SERIALIZE
+    (
+        READWRITE(prevout);
+        READWRITE(scriptSig);
+        READWRITE(nSequence);
+    )
+
+    bool IsFinal() const
+    {
+        return (nSequence == std::numeric_limits<unsigned int>::max());
+    }
+
+    friend bool operator==(const CTxIn& a, const CTxIn& b)
+    {
+        return (a.prevout   == b.prevout &&
+                a.scriptSig == b.scriptSig &&
+                a.nSequence == b.nSequence);
+    }
+
+    friend bool operator!=(const CTxIn& a, const CTxIn& b)
     {
         return !(a == b);
     }
@@ -277,6 +261,274 @@ public:
     void print() const;
 };
 
+class CAuxPow;
+class CBlock;
+
+
+
+
+struct CDiskBlockPos
+{
+    int nFile;
+    unsigned int nPos;
+
+    IMPLEMENT_SERIALIZE(
+        READWRITE(VARINT(nFile));
+        READWRITE(VARINT(nPos));
+    )
+
+    CDiskBlockPos() {
+        SetNull();
+    }
+
+  CDiskBlockPos(int nFileIn, unsigned int nPosIn) {
+    nFile = nFileIn;
+    nPos = nPosIn;
+  }
+
+  friend bool operator==(const CDiskBlockPos &a, const CDiskBlockPos &b) {
+        return (a.nFile == b.nFile && a.nPos == b.nPos);
+    }
+
+  friend bool operator!=(const CDiskBlockPos &a, const CDiskBlockPos &b) {
+        return !(a == b);
+    }
+
+  void SetNull() { nFile = -1; nPos = 0; }
+  bool IsNull() const { return (nFile == -1); }
+  
+};
+
+class CBlockIndex;
+
+/** A transaction with a merkle branch linking it to the block chain. */
+class CMerkleTx : public CTransaction
+{
+private:
+    int GetDepthInMainChainINTERNAL(CBlockIndex* &pindexRet) const;
+
+public:
+    uint256 hashBlock;
+    std::vector<uint256> vMerkleBranch;
+    int nIndex;
+
+    // memory only
+    mutable bool fMerkleVerified;
+
+
+    CMerkleTx()
+    {
+        Init();
+    }
+
+    CMerkleTx(const CTransaction& txIn) : CTransaction(txIn)
+    {
+        Init();
+    }
+
+    void Init()
+    {
+        hashBlock = 0;
+        nIndex = -1;
+        fMerkleVerified = false;
+    }
+
+
+    IMPLEMENT_SERIALIZE
+    (
+        nSerSize += SerReadWrite(s, *(CTransaction*)this, nType, nVersion, ser_action);
+        nVersion = this->nVersion;
+        READWRITE(hashBlock);
+        READWRITE(vMerkleBranch);
+        READWRITE(nIndex);
+    )
+
+
+    int SetMerkleBranch(const CBlock* pblock=NULL);
+
+    // Return depth of transaction in blockchain:
+    // -1  : not in blockchain, and not in memory pool (conflicted transaction)
+    //  0  : in memory pool, waiting to be included in a block
+    // >=1 : this many blocks deep in the main chain
+    int GetDepthInMainChain(CBlockIndex* &pindexRet) const;
+    int GetDepthInMainChain() const { CBlockIndex *pindexRet; return GetDepthInMainChain(pindexRet); }
+    bool IsInMainChain() const { CBlockIndex *pindexRet; return GetDepthInMainChainINTERNAL(pindexRet) > 0; }
+    int GetBlocksToMaturity() const;
+    bool AcceptToMemoryPool(bool fLimitFree=true);
+};
+
+/** Header for merge-mining data in the coinbase.  */
+static const unsigned char pchMergedMiningHeader[] = {0xfa, 0xbe, 'm', 'm'};
+
+/**
+ * Data for the merge-mining auxpow.  This is a merkle tx (the parent block's
+ * coinbase tx) that can be verified to be in the parent block, and this
+ * transaction's input (the coinbase script) contains the reference
+ * to the actual merge-mined block.
+ */
+class CAuxPow : public CMerkleTx
+{
+    /* Public for the unit tests.  */
+public:
+    /** The merkle branch connecting the aux block to our coinbase.  */
+    std::vector<uint256> vChainMerkleBranch;
+
+    /** Merkle tree index of the aux block header in the coinbase.  */
+    int nChainIndex;
+
+    /** Parent block header (on which the real PoW is done).  */
+    CPureBlockHeader parentBlock;
+
+public:
+    /* Prevent accidental conversion.  */
+    inline explicit CAuxPow(const CTransaction& txIn)
+        : CMerkleTx(txIn)
+    {
+    }
+
+    inline CAuxPow()
+        : CMerkleTx()
+    {
+    }
+
+    IMPLEMENT_SERIALIZE
+      (
+       nSerSize += SerReadWrite(s, *(CMerkleTx*)this, nType, nVersion,ser_action);
+       nVersion = this->nVersion;
+       READWRITE(vChainMerkleBranch);
+       READWRITE(nChainIndex);
+       READWRITE(parentBlock);
+       )
+    
+    /**
+   * Check the auxpow, given the merge-mined block's hash and our chain ID.
+   * Note that this does not verify the actual PoW on the parent block!  It
+   * just confirms that all the merkle branches are valid.
+   * @param hashAuxBlock Hash of the merge-mined block.
+   * @param nChainId The auxpow chain ID of the block to check.
+   * @param params Consensus parameters.
+   * @return True if the auxpow is valid.
+   */
+    bool check(const uint256& hashAuxBlock, int nChainId, const CChainParams& params) const;
+
+    /**
+   * Get the parent block's hash.  This is used to verify that it
+   * satisfies the PoW requirement.
+   * @return The parent block hash.
+   */
+    inline uint256
+    getParentBlockPoWHash(int algo) const
+    {
+        return parentBlock.GetPoWHash(algo);
+    }
+
+    /**
+   * Return parent block.  This is only used for the temporary parentblock
+   * auxpow version check.
+   * @return The parent block.
+   */
+    /* FIXME: Remove after the hardfork.  */
+    inline const CPureBlockHeader&
+    getParentBlock() const
+    {
+        return parentBlock;
+    }
+
+    /**
+   * Calculate the expected index in the merkle tree.  This is also used
+   * for the test-suite.
+   * @param nNonce The coinbase's nonce value.
+   * @param nChainId The chain ID.
+   * @param h The merkle block height.
+   * @return The expected index for the aux hash.
+   */
+    static int getExpectedIndex(int nNonce, int nChainId, unsigned h);
+};
+
+/** Nodes collect new transactions into a block, hash them into a hash tree,
+ * and scan through nonce values to make the block's hash satisfy proof-of-work
+ * requirements.  When they solve the proof-of-work, they broadcast the block
+ * to everyone and the block is added to the block chain.  The first transaction
+ * in the block is a special one that creates a new coin owned by the creator
+ * of the block.
+ */
+class CBlockHeader : public CPureBlockHeader
+{
+public:
+    boost::shared_ptr<CAuxPow> auxpow;
+
+    CBlockHeader()
+    {
+        SetNull();
+    }
+
+    unsigned int GetSerializeSize(int nType, int nVersion) const	\
+    {                                           \
+        CSerActionGetSerializeSize ser_action;  \
+        const bool fGetSize = true;             \
+        const bool fWrite = false;              \
+        const bool fRead = false;               \
+        unsigned int nSerSize = 0;              \
+        ser_streamplaceholder s;                \
+        assert(fGetSize||fWrite||fRead); /* suppress warning */ \
+        s.nType = nType;                        \
+        s.nVersion = nVersion;                  \
+	READWRITE(*(CPureBlockHeader*)this);
+	nVersion = this->nVersion;
+        return nSerSize;                        \
+    }                                           \
+    template<typename Stream>                   \
+    void Serialize(Stream& s, int nType, int nVersion) const \
+    {                                           \
+        CSerActionSerialize ser_action;         \
+        const bool fGetSize = false;            \
+        const bool fWrite = true;               \
+        const bool fRead = false;               \
+        unsigned int nSerSize = 0;              \
+        assert(fGetSize||fWrite||fRead); /* suppress warning */ \
+	READWRITE(*(CPureBlockHeader*)this);
+	nVersion = this->nVersion;
+    }                                           \
+    template<typename Stream>                   \
+    void Unserialize(Stream& s, int nType, int nVersion)  \
+    {                                           \
+        CSerActionUnserialize ser_action;       \
+        const bool fGetSize = false;            \
+        const bool fWrite = false;              \
+        const bool fRead = true;                \
+        unsigned int nSerSize = 0;              \
+        assert(fGetSize||fWrite||fRead); /* suppress warning */ \
+	READWRITE(*(CPureBlockHeader*)this);
+	nVersion = this->nVersion;
+	if (this->IsAuxpow()) {
+	  auxpow.reset(new CAuxPow());
+	  assert(auxpow);
+	  READWRITE(*auxpow);
+	} else {
+	  auxpow.reset();
+	}
+    }
+
+    void SetNull()
+    {
+      CPureBlockHeader::SetNull();
+      auxpow.reset();
+    }
+    
+    void SetAuxpow(CAuxPow* apow)
+    {
+      if (apow)
+	{
+	  auxpow.reset(apow);
+	  CPureBlockHeader::SetAuxpow(true);
+	} else
+	{
+	  auxpow.reset();
+	  CPureBlockHeader::SetAuxpow(false);
+	}
+    }
+};
+
 /** wrapper for CTxOut that provides a more compact serialization */
 class CTxOutCompressor
 {
@@ -358,112 +610,6 @@ public:
     )
 };
 
-
-/** Nodes collect new transactions into a block, hash them into a hash tree,
- * and scan through nonce values to make the block's hash satisfy proof-of-work
- * requirements.  When they solve the proof-of-work, they broadcast the block
- * to everyone and the block is added to the block chain.  The first transaction
- * in the block is a special one that creates a new coin owned by the creator
- * of the block.
- */
-class CBlockHeader
-{
-public:
-    // header
-    static const int CURRENT_VERSION=2;
-    int nVersion;
-    uint256 hashPrevBlock;
-    uint256 hashMerkleRoot;
-    unsigned int nTime;
-    unsigned int nBits;
-    unsigned int nNonce;
-
-    CBlockHeader()
-    {
-        SetNull();
-    }
-
-    IMPLEMENT_SERIALIZE
-    (
-        READWRITE(this->nVersion);
-        nVersion = this->nVersion;
-        READWRITE(hashPrevBlock);
-        READWRITE(hashMerkleRoot);
-        READWRITE(nTime);
-        READWRITE(nBits);
-        READWRITE(nNonce);
-    )
-
-    void SetNull()
-    {
-        nVersion = CBlockHeader::CURRENT_VERSION;
-        hashPrevBlock = 0;
-        hashMerkleRoot = 0;
-        nTime = 0;
-        nBits = 0;
-        nNonce = 0;
-    }
-
-    void SetAlgo(int algo)
-    {
-      switch(algo)
-	{
-	case ALGO_SHA256D:
-	  nVersion |= BLOCK_VERSION_SHA256D;
-	  break;
-	case ALGO_SCRYPT:
-	  nVersion |= BLOCK_VERSION_SCRYPT;
-	  break;
-	case ALGO_ARGON2:
-	  nVersion |= BLOCK_VERSION_ARGON2;
-	  break;
-	case ALGO_X17:
-	  nVersion |= BLOCK_VERSION_X17;
-	  break;
-	case ALGO_LYRA2REv2:
-	  nVersion |= BLOCK_VERSION_LYRA2REv2;
-	  break;
-	default:
-	  break;
-	}
-    }
-
-    int GetAlgo () {
-      switch (nVersion & BLOCK_VERSION_ALGO)
-	{
-	case BLOCK_VERSION_SHA256D:
-	  return ALGO_SHA256D;
-	case BLOCK_VERSION_SCRYPT:
-	  return ALGO_SCRYPT;
-	case BLOCK_VERSION_ARGON2:
-	  return ALGO_ARGON2;
-	case BLOCK_VERSION_X17:
-	  return ALGO_X17;
-	case BLOCK_VERSION_LYRA2REv2:
-	  return ALGO_LYRA2REv2;
-	}
-      return ALGO_SCRYPT;
-    }
-
-    void SetUpdateSSF()
-    {
-      nVersion |= BLOCK_VERSION_UPDATE_SSF;
-    }
-    
-    bool IsNull() const
-    {
-        return (nBits == 0);
-    }
-
-    uint256 GetHash() const;
-
-    int64_t GetBlockTime() const
-    {
-        return (int64_t)nTime;
-    }
-};
-
-
 class CBlock : public CBlockHeader
 {
 public:
@@ -509,42 +655,6 @@ public:
         return block;
     }
 
-    uint256 GetPoWHash(int algo) const
-    {
-      switch(algo) {
-      case ALGO_SHA256D:
-	return GetHash();
-      case ALGO_SCRYPT:
-	{
-	  uint256 thash;
-	  hash_scrypt(BEGIN(nVersion),BEGIN(thash));
-	  return thash;
-	}
-      case ALGO_ARGON2:
-	{
-	  uint256 thash;
-	  hash_argon2(BEGIN(nVersion),BEGIN(thash));
-	  return thash;
-	}
-      case ALGO_X17:
-	return hash_x17(BEGIN(nVersion), END(nNonce));
-      case ALGO_LYRA2REv2:
-	{
-	  uint256 thash;
-	  hash_lyra2rev2(BEGIN(nVersion),BEGIN(thash));
-	  return thash;
-	}
-      }
-      uint256 thash;
-      hash_scrypt(BEGIN(nVersion),BEGIN(thash));
-      return thash;
-    }
-	    
-    uint256 GetPoWHash() const
-    {
-      return GetPoWHash(((CBlockHeader*)this)->GetAlgo());
-    }
-
     uint256 BuildMerkleTree() const;
 
     const uint256 &GetTxHash(unsigned int nIndex) const {
@@ -558,11 +668,256 @@ public:
     void print() const;
 };
 
+enum BlockStatus {
+    BLOCK_VALID_UNKNOWN      =    0,
+    BLOCK_VALID_HEADER       =    1, // parsed, version ok, hash satisfies claimed PoW, 1 <= vtx count <= max, timestamp not in future
+    BLOCK_VALID_TREE         =    2, // parent found, difficulty matches, timestamp >= median previous, checkpoint
+    BLOCK_VALID_TRANSACTIONS =    3, // only first tx is coinbase, 2 <= coinbase input script length <= 100, transactions valid, no duplicate txids, sigops, size, merkle root
+    BLOCK_VALID_CHAIN        =    4, // outputs do not overspend inputs, no double spends, coinbase output ok, immature coinbase spends, BIP30
+    BLOCK_VALID_SCRIPTS      =    5, // scripts/signatures ok
+    BLOCK_VALID_MASK         =    7,
+
+    BLOCK_HAVE_DATA          =    8, // full block available in blk*.dat
+    BLOCK_HAVE_UNDO          =   16, // undo data available in rev*.dat
+    BLOCK_HAVE_MASK          =   24,
+
+    BLOCK_FAILED_VALID       =   32, // stage after last reached validness failed
+    BLOCK_FAILED_CHILD       =   64, // descends from failed block
+    BLOCK_FAILED_MASK        =   96
+};
+
+FILE* OpenDiskFile(const CDiskBlockPos &pos, const char *prefix, bool fReadOnly);
+
+FILE* OpenBlockFile(const CDiskBlockPos &pos, bool fReadOnly = false);
+
+bool CheckAuxPowProofOfWork(const CBlockHeader& block, const CChainParams& params);
+
+/** The block chain is a tree shaped structure starting with the
+ * genesis block at the root, with each block potentially having multiple
+ * candidates to be the next block. A blockindex may have multiple pprev pointing
+ * to it, but at most one of them can be part of the currently active branch.
+ */
+class CBlockIndex
+{
+public:
+    // pointer to the hash of the block, if any. memory is owned by this CBlockIndex
+    const uint256* phashBlock;
+
+    // pointer to the index of the predecessor of this block
+    CBlockIndex* pprev;
+
+    // pointer to the AuxPoW header, if this block has one
+    boost::shared_ptr<CAuxPow> pauxpow;
+
+    // height of the entry in the chain. The genesis block has height 0
+    int nHeight;
+
+    // track the amount of coins emitted since genesis block, allowing us to determine max block reward
+    int64_t nMoneySupply;
+
+    // the scaling factor for the block
+    double subsidyScalingFactor;
+    
+    // Which # file this block is stored in (blk?????.dat)
+    int nFile;
+
+    // Byte offset within blk?????.dat where this block's data is stored
+    unsigned int nDataPos;
+
+    // Byte offset within rev?????.dat where this block's undo data is stored
+    unsigned int nUndoPos;
+
+    // (memory only) Total amount of work (expected number of hashes) in the chain up to and including this block
+    uint256 nChainWork;
+
+    // Number of transactions in this block.
+    // Note: in a potential headers-first mode, this number cannot be relied upon
+    unsigned int nTx;
+
+    // (memory only) Number of transactions in the chain up to and including this block
+    unsigned int nChainTx; // change to 64-bit type when necessary; won't happen before 2030
+
+    // Verification status of this block. See enum BlockStatus
+    unsigned int nStatus;
+
+    // block header
+    int nVersion;
+    uint256 hashMerkleRoot;
+    unsigned int nTime;
+    unsigned int nBits;
+    unsigned int nNonce;
+
+    // (memory only) Sequencial id assigned to distinguish order in which blocks are received.
+    uint32_t nSequenceId;
+
+    CBlockIndex()
+    {
+        phashBlock = NULL;
+        pprev = NULL;
+        nHeight = 0;
+        nMoneySupply = 0;
+	subsidyScalingFactor = 0;
+        nFile = 0;
+        nDataPos = 0;
+        nUndoPos = 0;
+        nChainWork = 0;
+        nTx = 0;
+        nChainTx = 0;
+        nStatus = 0;
+        nSequenceId = 0;
+
+        nVersion       = 0;
+        hashMerkleRoot = 0;
+        nTime          = 0;
+        nBits          = 0;
+        nNonce         = 0;
+    }
+
+    CBlockIndex(CBlockHeader& block)
+    {
+        phashBlock = NULL;
+        pprev = NULL;
+        nHeight = 0;
+        nMoneySupply = 0;
+	subsidyScalingFactor = 0;
+        nFile = 0;
+        nDataPos = 0;
+        nUndoPos = 0;
+        nChainWork = 0;
+        nTx = 0;
+        nChainTx = 0;
+        nStatus = 0;
+        nSequenceId = 0;
+
+        nVersion       = block.nVersion;
+        hashMerkleRoot = block.hashMerkleRoot;
+        nTime          = block.nTime;
+        nBits          = block.nBits;
+        nNonce         = block.nNonce;
+    }
+
+    CDiskBlockPos GetBlockPos() const {
+        CDiskBlockPos ret;
+        if (nStatus & BLOCK_HAVE_DATA) {
+            ret.nFile = nFile;
+            ret.nPos  = nDataPos;
+        }
+        return ret;
+    }
+
+    CDiskBlockPos GetUndoPos() const {
+        CDiskBlockPos ret;
+        if (nStatus & BLOCK_HAVE_UNDO) {
+            ret.nFile = nFile;
+            ret.nPos  = nUndoPos;
+        }
+        return ret;
+    }
+
+    CBlockHeader GetBlockHeader() const
+    {
+        CBlockHeader block;
+	if (IsAuxpow())
+	  {
+	    const CDiskBlockPos pos = GetBlockPos();
+	    CAutoFile filein(OpenBlockFile(pos, true), SER_DISK, CLIENT_VERSION);
+	    if (filein==NULL)
+	      LogPrintf("ReadBlockFromDisk: OpenBlockFile failed for %s", "");
+	    try {
+	      filein >> block;
+	    }
+	    catch (const std::exception& e) {
+	      LogPrintf("%s: Deserialize or I/O error - %s at %s", __func__, e.what(), "");
+	    }
+	    if (!CheckAuxPowProofOfWork(block, Params()))
+	      LogPrintf("ReadBlockFromDisk: Errors in block header at %s", "");
+	    if (block.GetHash() != GetBlockHash())
+	      LogPrintf("ReadBlockFromDisk(CBlock&, CBlockIndex*): GetHash() doesn't match index for %s at %s",
+			   ToString(), "");
+	    return block;
+	  }
+        block.nVersion       = nVersion;
+        if (pprev)
+            block.hashPrevBlock = pprev->GetBlockHash();
+        block.hashMerkleRoot = hashMerkleRoot;
+        block.nTime          = nTime;
+        block.nBits          = nBits;
+        block.nNonce         = nNonce;
+        return block;
+    }
+
+    uint256 GetBlockHash() const
+    {
+        return *phashBlock;
+    }
+
+    int64_t GetBlockTime() const
+    {
+        return (int64_t)nTime;
+    }
+
+    CBigNum GetBlockWork() const
+    {
+        CBigNum bnTarget;
+        bnTarget.SetCompact(nBits);
+        if (bnTarget <= 0)
+            return 0;
+        return (CBigNum(1)<<256) / (bnTarget+1);
+    }
+
+    bool CheckIndex() const
+    {
+        return true;
+    }
+
+    enum { nMedianTimeSpan=11 };
+
+    int64_t GetMedianTimePast() const
+    {
+        int64_t pmedian[nMedianTimeSpan];
+        int64_t* pbegin = &pmedian[nMedianTimeSpan];
+        int64_t* pend = &pmedian[nMedianTimeSpan];
+
+        const CBlockIndex* pindex = this;
+        for (int i = 0; i < nMedianTimeSpan && pindex; i++, pindex = pindex->pprev)
+            *(--pbegin) = pindex->GetBlockTime();
+
+        std::sort(pbegin, pend);
+        return pbegin[(pend - pbegin)/2];
+    }
+
+    int64_t GetMedianTime() const;
+
+    /**
+     * Returns true if there are nRequired or more blocks of minVersion or above
+     * in the last nToCheck blocks, starting at pstart and going backwards.
+     */
+    static bool IsSuperMajority(int minVersion, const CBlockIndex* pstart,
+                                unsigned int nRequired, unsigned int nToCheck);
+
+    std::string ToString() const
+    {
+        return strprintf("CBlockIndex(pprev=%p, nHeight=%d, merkle=%s, hashBlock=%s)",
+            pprev, nHeight,
+            hashMerkleRoot.ToString().c_str(),
+            GetBlockHash().ToString().c_str());
+    }
+
+    void print() const
+    {
+        LogPrintf("%s\n", ToString().c_str());
+    }
+
+    inline bool IsAuxpow() const
+    {
+      return nVersion & BLOCK_VERSION_AUXPOW;
+    }
+};
 
 /** Describes a place in the block chain to another node such that if the
  * other node doesn't have the same branch, it can find a recent common trunk.
  * The further back it is, the further before the fork it may be.
- */
++ */
 struct CBlockLocator
 {
     std::vector<uint256> vHave;
@@ -591,5 +946,70 @@ struct CBlockLocator
         return vHave.empty();
     }
 };
+
+/** An in-memory indexed chain of blocks. */
+class CChain {
+private:
+    std::vector<CBlockIndex*> vChain;
+
+public:
+    /** Returns the index entry for the genesis block of this chain, or NULL if none. */
+    CBlockIndex *Genesis() const {
+        return vChain.size() > 0 ? vChain[0] : NULL;
+    }
+
+    /** Returns the index entry for the tip of this chain, or NULL if none. */
+    CBlockIndex *Tip() const {
+        return vChain.size() > 0 ? vChain[vChain.size() - 1] : NULL;
+    }
+
+    /** Returns the index entry at a particular height in this chain, or NULL if no such height exists. */
+    CBlockIndex *operator[](int nHeight) const {
+        if (nHeight < 0 || nHeight >= (int)vChain.size())
+            return NULL;
+        return vChain[nHeight];
+    }
+
+    /** Compare two chains efficiently. */
+    friend bool operator==(const CChain &a, const CChain &b) {
+        return a.vChain.size() == b.vChain.size() &&
+               a.vChain[a.vChain.size() - 1] == b.vChain[b.vChain.size() - 1];
+    }
+
+    /** Efficiently check whether a block is present in this chain. */
+    bool Contains(const CBlockIndex *pindex) const {
+        return (*this)[pindex->nHeight] == pindex;
+    }
+
+    /** Find the successor of a block in this chain, or NULL if the given index is not found or is the tip. */
+    CBlockIndex *Next(const CBlockIndex *pindex) const {
+        if (Contains(pindex))
+            return (*this)[pindex->nHeight + 1];
+        else
+            return NULL;
+    }
+
+    /** Return the maximal height in the chain. Is equal to chain.Tip() ? chain.Tip()->nHeight : -1. */
+    int Height() const {
+        return vChain.size() - 1;
+    }
+
+    /** Set/initialize a chain with a given tip. Returns the forking point. */
+    CBlockIndex *SetTip(CBlockIndex *pindex);
+
+    /** Return a CBlockLocator that refers to a block in this chain (by default the tip). */
+    CBlockLocator GetLocator(const CBlockIndex *pindex = NULL) const;
+
+    /** Find the last common block between this chain and a locator. */
+    CBlockIndex *FindFork(const CBlockLocator &locator) const;
+};
+
+/** The currently-connected chain of blocks. */
+extern CChain chainActive;
+
+//#include "coins.h"
+
+/** Global variable that points to the active CCoinsView (protected by cs_main) */
+//extern CCoinsViewCache *pcoinsTip;
 
 #endif
